+++ koubatad/src/MainMenu.hpp
@@ -39,7 +39,7 @@ private:
      */
     void showLeaderboard();
 
-    void takeAction(size_t currSelect) override;
+    void takeAction(int currSelect) override;
 };
