+++ koubatad/src/OptionsMenu.hpp
@@ -43,5 +43,5 @@ private:
      */
     void displaySettings();
 
-    void takeAction(size_t currSelect) override;
+    void takeAction(int currSelect) override;
 };
